more reliable focus management for controlled component
previously, in order for focus management to work, we assumed the parent
would rerender the tabs component when its onSelect handler was invoked.
that was probably a pretty safe assumption, but we no longer rely on it.

also made it more explicit (and enforced) that `selectedIndex` and
`onSelect` come as a pair. and improved test coverage around this stuff.